password validation, poppins and opens sans font, swet alert ,etc modify and responsive some

+++ src/pages/ManageMyFoods.jsx
@@ -3,13 +3,14 @@ import { useContext, useEffect, useState } from "react";
 import { Helmet } from "react-helmet";
 import { AuthContext } from "../provider/AuthProvider";
 import { Link } from "react-router-dom";
+import Swal from "sweetalert2";
 
 const ManageMyFoods = () => {
     const { user } = useContext(AuthContext)
     const [foods, setMyFoods] = useState([])
     useEffect(() => {
         const getData = async () => {
-            const { data } = await axios(`${import.meta.env.VITE_URL}/allFood/${user.email}`,{withCredentials: true})
+            const { data } = await axios(`${import.meta.env.VITE_URL}/allFood/${user.email}`, { withCredentials: true })
             console.log(data)
             setMyFoods(data)
         }
@@ -17,16 +18,34 @@ const ManageMyFoods = () => {
     }, [user])
 
     const handleDeleteFood = _id => {
-        console.log('delete', _id)
-        axios.delete(`${import.meta.env.VITE_URL}/allFoodsdelete/${_id}`)
-            .then(res => {
-                console.log(res.data)
-                if (res.data.deletedCount > 0) {
-                    const remeningFood = foods.filter(food => food._id !== _id);
-                    setMyFoods(remeningFood)
-                }
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, delete it!"
+        }).then((result) => {
+            if (result.isConfirmed) {
+                console.log('delete', _id)
+                axios.delete(`${import.meta.env.VITE_URL}/allFoodsdelete/${_id}`)
+                    .then(res => {
+                        console.log(res.data)
+                        if (res.data.deletedCount > 0) {
+                            const remeningFood = foods.filter(food => food._id !== _id);
+                            setMyFoods(remeningFood)
+                            Swal.fire({
+                                title: "Deleted!",
+                                text: "Your file has been deleted.",
+                                icon: "success"
+                            });
+                        }
 
-            })
+                    })
+            }
+        });
+
     }
     return (
         <div>
@@ -40,7 +59,7 @@ const ManageMyFoods = () => {
                     <table className="table">
                         {/* head */}
                         <thead>
-                            <tr className=" text-xl font-bold">
+                            <tr className=" bg-gray-200 text-xl font-bold">
                                 <th>No</th>
                                 <th>Image</th>
                                 <th>Food Name</th>
@@ -57,9 +76,9 @@ const ManageMyFoods = () => {
                                     <td><img className="w-[70px] rounded-lg" src={food.foodImage} alt="" /></td>
                                     <td>{food.foodName}</td>
                                     <td>{food.foodQuantity}</td>
-                                    <td>{food.foodStatus}</td>
+                                    <td > <button className={" btn text-white " + (food.foodStatus === 'available' ? 'bg-green-500' : 'bg-red-500')}>{food.foodStatus}</button></td>
                                     <td> <Link to={`/updatedPages/${food._id}`}>
-                                    <button className="btn btn-info hover:btn-ghost">Update</button>
+                                        <button className="btn btn-info hover:btn-ghost">Update</button>
                                     </Link></td>
                                     <td><button onClick={() => handleDeleteFood(food._id)} className="btn btn-error hover:btn-ghost">X</button></td>
                                 </tr>)
